Доска объявлений - работа с почтой

+++ bulletin_board/board/views.py
@@ -1,34 +1,31 @@
 from urllib.parse import urlencode
 
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.models import User
 from django.core.exceptions import ObjectDoesNotExist
-from django.db.models import Count
+from django.core.paginator import Paginator
+from django.db.models import Count, Q
 from django.http import HttpResponse, QueryDict, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.urls import reverse, reverse_lazy
 from django.views import generic
-from django.views.generic import ListView, DetailView, TemplateView, CreateView, UpdateView
+from django.views.generic import ListView, DetailView, TemplateView, CreateView, UpdateView, DeleteView
 
-from .filters import EventFilter, AuthorFilter
+from .filters import EventFilter, AuthorFilter, ResponseFilter
 from .models import Event, Responses, Players, Category, AnswertoResponse
 from .forms import AddPostForm, AddResponseForm
 from .utils import DataMixin
 
 # Create your views here.
 
-menu = [{'title' : 'Bulletin Board', 'URL' : 'main'},
-        {'title' : 'О сайте', 'URL' : 'about'},
-        {'title' : 'Создать объявление', 'URL' : 'create_event'},
-        {'title' : 'Найти объявление', 'URL' : 'search_event'},
-        {'title' : 'Найти автора', 'URL' : 'search_player'},
-        {'title' : 'Войти', 'URL' : 'login'},
-        {'title' : 'Создать профиль', 'URL' : 'create profile'}]
+
 
 
 
 class MainPage(DataMixin,ListView):
     context_object_name = 'events'
     template_name = 'index.html'
-    paginate_by = 1
 
     def get_queryset(self):
         return Event.objects.filter(is_published=True).annotate(total=Count('resp')).order_by('-time_create').select_related('player').prefetch_related('category')
@@ -41,7 +38,6 @@ class MainPage(DataMixin,ListView):
 class PostSearch(DataMixin,ListView):
     template_name = 'search_event.html'
     context_object_name = 'events'
-    paginate_by = 1
 
     def get_queryset(self):
         queryset = Event.published.all().annotate(total=Count('resp')).order_by('-time_create').select_related('player').prefetch_related('category')
@@ -69,32 +65,22 @@ class About(DataMixin,TemplateView):
 def show_by_category(request, category_slug):
     category = get_object_or_404(Category, slug=category_slug)
 
+
     author = True if request.path != reverse('category', kwargs={'category_slug' : category_slug}) else False
 
-    return render(request=request, template_name='category.html', context={'category': category, 'menu' : menu, 'author' : author})
+    page_number = request.GET.get('page')
 
-#
-# class Show_posts_by_category(DataMixin,ListView):
-#     template_name = 'category.html'
-#     context_object_name = 'category'
-#     paginate_by = 1
-#
-#     def get_queryset(self):
-#         return Category.objects.filter(slug=self.kwargs['category_slug'])[0]
-#
-#     def get_context_data(self, **kwargs):
-#         context = super().get_context_data(**kwargs)
-#         return self.get_mixin_objects(context,
-#                                       author=False)
+    return render(request=request, template_name='category.html', context={'category': category,  'author' : author, 'page_number': page_number})
 
 
 
 
 
-class ShowPost(DataMixin,DetailView):
+class ShowPost(LoginRequiredMixin,DataMixin,DetailView):
     template_name = 'detail_event.html'
     slug_url_kwarg = 'event_slug'
     context_object_name = 'event'
+
 
     def get_object(self, queryset=None):
         try:
@@ -117,19 +103,28 @@ class ShowPost(DataMixin,DetailView):
 
 
 
-class CreateEvent(DataMixin,CreateView):
+class CreateEvent(LoginRequiredMixin,DataMixin,CreateView):
     template_name = 'create_event.html'
-    model = Responses
     form_class = AddPostForm
 
     def form_valid(self, form):
-        form.save()
+        f = form.save(commit=False)
+        f.player = Players.objects.get(user_id=self.request.user.id)
         return super().form_valid(form)
+
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        return self.get_mixin_objects(context, title=f'Создание объявления')
+        return self.get_mixin_objects(context, title=f'Создание объявления', update=False, player=Players.objects.get(user_id=self.request.user.id))
 
-class UpdateEvent(DataMixin,UpdateView):
+    def dispatch(self, request, *args, **kwargs):
+        if Players.objects.filter(user_id=self.request.user.id).exists() or self.request.user.is_superuser:
+            return super(CreateEvent, self).dispatch(request, *args, **kwargs)
+
+        return redirect('create profile')
+
+
+
+class UpdateEvent(LoginRequiredMixin,DataMixin,UpdateView):
     model = Event
     form_class = AddPostForm
     template_name = 'create_event.html'
@@ -139,9 +134,16 @@ class UpdateEvent(DataMixin,UpdateView):
         context = super().get_context_data()
         return self.get_mixin_objects(context,
                                       title=f'Редактирование объявления',
-                                      update=True)
+                                      update=True,
+                                      player=Players.objects.get(user_id=self.request.user.id))
 
-class CreateResponse(DataMixin,CreateView):
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user == Event.objects.get(pk=self.kwargs[self.pk_url_kwarg]).player.user or self.request.user.is_superuser:
+            return super(UpdateEvent, self).dispatch(request, *args, **kwargs)
+
+        return redirect(reverse('detail',kwargs={'event_slug': Event.objects.get(pk=self.kwargs[self.pk_url_kwarg]).slug}))
+
+class CreateResponse(LoginRequiredMixin,DataMixin,CreateView):
     template_name = 'create_response.html'
     model = Responses
     form_class = AddResponseForm
@@ -150,17 +152,25 @@ class CreateResponse(DataMixin,CreateView):
     def form_valid(self, form):
         resp = form.save(commit=False)
         resp.event = Event.objects.get(pk=self.kwargs[self.pk_url_kwarg])
+        resp.player = Players.objects.get(user_id=self.request.user.id)
         return super().form_valid(form)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         return self.get_mixin_objects(context,
                                       responses=Responses.objects.filter(event_id=self.kwargs[self.pk_url_kwarg]),
-                                      event=Event.objects.get(pk=self.kwargs[self.pk_url_kwarg]))
+                                      event=Event.objects.get(pk=self.kwargs[self.pk_url_kwarg]),
+                                      player=Players.objects.get(user_id=self.request.user.id))
+
+    def dispatch(self, request, *args, **kwargs):
+        if Players.objects.filter(user_id=self.request.user.id).exists() or self.request.user.is_superuser:
+            return super(CreateResponse, self).dispatch(request,*args,**kwargs)
+
+        return redirect('create profile')
 
 
 
-class UpdateResponse(DataMixin,UpdateView):
+class UpdateResponse(LoginRequiredMixin,DataMixin,UpdateView):
     model = Responses
     template_name = 'update_response.html'
     fields = ['message']
@@ -180,29 +190,37 @@ class UpdateResponse(DataMixin,UpdateView):
                                       event=event,
                                       responses=Responses.objects.filter(event_id=event.pk),
                                       ourresp=Responses.objects.get(pk=self.kwargs[self.pk_url_kwarg]))
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user == Responses.objects.get(pk=self.kwargs[self.pk_url_kwarg]).player.user or self.request.user.is_superuser:
+            return super(UpdateResponse, self).dispatch(request,*args,**kwargs)
+
+        return redirect(reverse('response', kwargs={'event_id' : Responses.objects.get(pk=self.kwargs[self.pk_url_kwarg]).event.id}))
 
 
+
+
+
+
+@login_required
 def delete_post(request, event_slug):
     event = get_object_or_404(Event, slug=event_slug)
 
     data = {
         'title' : f'Вы действительно хотите удалить объявление "{event.title}" ?',
-        'menu' : menu,
         'event' : event
 
     }
 
+    return render(request, 'delete_post.html', data) if request.user == event.player.user or request.user.is_superuser else redirect(reverse('detail', kwargs={'event_slug': event_slug}))
 
-    return render(request, 'delete_post.html', data)
-
-
+@login_required
 def successful_delete(request, event_slug):
     event = get_object_or_404(Event, slug=event_slug)
-    Event.objects.get(slug=event_slug).delete()
-    return render(request, 'successful_delete.html', context={'event' : event, 'menu' : menu})
+    event.delete()
+    return render(request, 'successful_delete.html', context={'event' : event})
 
 
-class PlayerProfile(DataMixin,DetailView):
+class PlayerProfile(LoginRequiredMixin, DataMixin,DetailView):
     template_name = 'player.html'
     pk_url_kwarg = 'player_id'
     context_object_name = 'player'
@@ -225,19 +243,32 @@ class PlayerProfile(DataMixin,DetailView):
 
 
 
-class CreateProfile(DataMixin,CreateView):
+class CreateProfile(LoginRequiredMixin, DataMixin,CreateView):
     model = Players
-    fields = '__all__'
+    fields = ['name', 'age', 'sex', 'city','level','guild', 'profile_picture']
     template_name = 'create_profile.html'
 
+
+    def form_valid(self, form):
+        f = form.save(commit=False)
+        f.user = self.request.user
+        return super().form_valid(form)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         return self.get_mixin_objects(context)
 
-class UpdateProfile(DataMixin,UpdateView):
+    def dispatch(self, request, *args, **kwargs):
+        if Players.objects.filter(user_id=self.request.user.id).exists():
+            return redirect(reverse('player', kwargs={'player_id': Players.objects.get(user_id=self.request.user.id).id}))
+
+        return super(CreateProfile, self).dispatch(request, *args, **kwargs)
+
+
+
+class UpdateProfile(LoginRequiredMixin, DataMixin,UpdateView):
     model = Players
-    fields = '__all__'
+    fields = ['name', 'age', 'sex', 'city','level','guild', 'profile_picture']
     pk_url_kwarg = 'player_id'
     template_name = 'create_profile.html'
 
@@ -246,11 +277,34 @@ class UpdateProfile(DataMixin,UpdateView):
         return self.get_mixin_objects(context,
                                       title=Players.objects.get(pk=self.kwargs[self.pk_url_kwarg]))
 
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user == Players.objects.get(pk=self.kwargs[self.pk_url_kwarg]).user or self.request.is_superuser:
+            return super(UpdateProfile, self).dispatch(request, *args, **kwargs)
 
-class SearchProfile(DataMixin,ListView):
+
+        return redirect(reverse('player', kwargs={'player_id':self.kwargs[self.pk_url_kwarg]}))
+
+
+
+
+class DeleteProfile(LoginRequiredMixin, DeleteView):
+    template_name = 'delete_profile.html'
+    success_url = reverse_lazy('main')
+    model = Players
+    pk_url_kwarg = 'player_id'
+    context_object_name = 'author'
+
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user == Players.objects.get(pk=self.kwargs[self.pk_url_kwarg]).user or self.request.is_superuser:
+            return super(DeleteProfile, self).dispatch(request, *args, **kwargs)
+
+        return redirect(reverse('player', kwargs={'player_id': self.kwargs[self.pk_url_kwarg]}))
+
+
+class SearchProfile(LoginRequiredMixin,DataMixin,ListView):
     template_name = 'search_player.html'
     context_object_name = 'authors'
-    paginate_by = 1
+
 
     def get_queryset(self):
         queryset = Players.objects.all().select_related('guild').select_related('user').prefetch_related('events')
@@ -266,7 +320,7 @@ class SearchProfile(DataMixin,ListView):
 
 
 
-class AllResponses(DataMixin,ListView):
+class AllResponses(LoginRequiredMixin,ListView):
     template_name = 'responses.html'
     context_object_name = 'responses'
 
@@ -276,19 +330,25 @@ class AllResponses(DataMixin,ListView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        return self.get_mixin_objects(context,event=Event.objects.filter(pk=self.kwargs['event_id']).prefetch_related('category')[0])
+        context['event'] = Event.objects.filter(pk=self.kwargs['event_id']).prefetch_related('category')[0]
+        return context
 
+@login_required
 def delete_response(request,resp_id):
     resp = get_object_or_404(Responses, pk=resp_id)
-    Responses.objects.get(pk=resp_id).delete()
+
+    if request.user == Responses.objects.get(pk=resp_id).player.user or request.user.is_superuser or Responses.objects.get(pk=resp_id).event.player.user:
+        Responses.objects.get(pk=resp_id).delete()
+
 
     return redirect(reverse('response', kwargs={'event_id': resp.event.pk}))
 
-class AllAuthors(DataMixin,ListView):
+
+
+class AllAuthors(LoginRequiredMixin,DataMixin,ListView):
     model = Players
     template_name = 'authors.html'
     context_object_name = 'authors'
-    paginate_by = 1
 
     def get_queryset(self):
         return Players.objects.all().select_related('guild').prefetch_related('events')
@@ -298,7 +358,7 @@ class AllAuthors(DataMixin,ListView):
         return self.get_mixin_objects(context, title='Все авторы')
 
 
-class AllAnswers(DataMixin,ListView):
+class AllAnswers(LoginRequiredMixin,ListView):
     template_name = 'show_answers_to_response.html'
     context_object_name = 'answers'
 
@@ -309,21 +369,22 @@ class AllAnswers(DataMixin,ListView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         event_id = Responses.objects.get(pk=self.kwargs['resp_id']).event.pk
-        return self.get_mixin_objects(context,
-                                      responses=Responses.objects.filter(event_id=event_id),
-                                      ourresp=Responses.objects.get(pk=self.kwargs['resp_id']).pk,
-                                      event=Responses.objects.get(pk=self.kwargs['resp_id']).event)
+        context['responses'] = Responses.objects.filter(event_id=event_id)
+        context['ourresp'] = Responses.objects.get(pk=self.kwargs['resp_id']).pk
+        context['event'] = Responses.objects.get(pk=self.kwargs['resp_id']).event
+        return context
 
 
-class CreateAnswerToResponse(DataMixin,CreateView):
+class CreateAnswerToResponse(LoginRequiredMixin,DataMixin,CreateView):
     model = AnswertoResponse
-    fields = ['player', 'message']
+    fields = ['message']
     template_name = 'create_answer_to_response.html'
 
 
     def form_valid(self, form):
         answer = form.save(commit=False)
         answer.resp = Responses.objects.get(pk=self.kwargs['resp_id'])
+        answer.player = Players.objects.get(user_id=self.request.user.id)
         return super().form_valid(form)
 
     def get_context_data(self, **kwargs):
@@ -332,10 +393,18 @@ class CreateAnswerToResponse(DataMixin,CreateView):
         return self.get_mixin_objects(context,
                                       responses=Responses.objects.filter(event_id=event_id),
                                       ourresp=Responses.objects.get(pk=self.kwargs['resp_id']).pk,
-                                      event=Responses.objects.get(pk=self.kwargs['resp_id']).event)
+                                      event=Responses.objects.get(pk=self.kwargs['resp_id']).event,
+                                      player = Players.objects.get(user_id=self.request.user.id))
+
+    def dispatch(self, request, *args, **kwargs):
+        if Players.objects.filter(user_id=self.request.user.id).exists():
+            return super(CreateAnswerToResponse,self).dispatch(request,*args,**kwargs)
+
+        return redirect(reverse('show answers to response', kwargs={'resp_id' : self.kwargs['resp_id']}))
 
 
-class UpdateAnswerToResponse(DataMixin,UpdateView):
+
+class UpdateAnswerToResponse(LoginRequiredMixin,DataMixin,UpdateView):
     model = AnswertoResponse
     fields = ['message']
     template_name = 'update_answer_to_response.html'
@@ -359,13 +428,60 @@ class UpdateAnswerToResponse(DataMixin,UpdateView):
                                       author = AnswertoResponse.objects.get(pk=self.kwargs[self.pk_url_kwarg]).player
                                       )
 
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user == AnswertoResponse.objects.get(pk=self.kwargs[self.pk_url_kwarg]).player.user \
+                or self.request.user.is_superuser:
+            return super(UpdateAnswerToResponse, self).dispatch(request,*args, **kwargs)
 
+        return redirect(reverse('show answers to response',
+                                kwargs={'resp_id' : AnswertoResponse.objects.get(pk=self.kwargs['answ_id']).resp.id}))
+
+@login_required
 def delete_answer(request, answ_id):
     answer = get_object_or_404(AnswertoResponse, pk=answ_id)
-    AnswertoResponse.objects.get(pk=answ_id).delete()
+
+    if request.user == AnswertoResponse.objects.get(pk=answ_id).player.user or request.user.is_superuser or AnswertoResponse.objects.get(pk=answ_id).resp.event.player.user:
+
+        AnswertoResponse.objects.get(pk=answ_id).delete()
 
     return redirect(reverse('show answers to response', kwargs={'resp_id': answer.resp_id}))
 
+class PrivateWebPage(LoginRequiredMixin, DataMixin,ListView):
+    template_name = 'private_page.html'
+    context_object_name = 'responses'
 
-def login(request):
-    return HttpResponse('Авторизация')
+    def get_queryset(self):
+        events = Event.published.filter(player_id=self.kwargs['player_id']).prefetch_related('resp').prefetch_related('category').select_related('player')
+        if len(events) > 0:
+            responses = Responses.objects.filter(Q(event__in=events) & ~Q(player_id=self.kwargs['player_id']))
+            self.filterset = ResponseFilter(self.request.GET, responses) if len(responses) > 0 else None
+            return self.filterset.qs if self.filterset is not None else None
+        else:
+            self.filterset = None
+            return None
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['filterset'] = self.filterset
+        return context
+
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user == Players.objects.get(pk=self.kwargs['player_id']).user \
+                or self.request.user.is_superuser:
+            return super(PrivateWebPage, self).dispatch(request,*args, **kwargs)
+
+        return redirect(reverse('player', kwargs={'player_id' : self.kwargs['player_id']}))
+
+def take_player_by_response(request, resp_id):
+    resp = get_object_or_404(Responses, pk=resp_id)
+
+    if not resp.is_taken and request.user == resp.event.player.user and resp.player and resp.event.player.user != resp.player.user:
+        Responses.objects.filter(pk=resp_id).update(is_taken=True)
+
+    return redirect(reverse('take player by response', kwargs={'player_id' : request.user.player.id}))\
+        if request.user == resp.event.player.user else redirect(reverse('response', kwargs={'event_id' : resp.event.id}))
+
+
+
+
+
